Refactor - extracted the Card model to a seperate file in a models folder and also extracted the interface definitions out in a seperate folder and file
