+++ src/models/card.ts
@@ -1,0 +1,16 @@
+import { CardDTO, Card, CardSummary } from "../types/index";
+
+const createCard = (card: CardDTO): Card => {
+  const getSummary = (): CardSummary => {
+    return {
+      title: card.title,
+      id: card.id,
+    };
+  };
+
+  return {
+    getSummary,
+  };
+};
+
+export default createCard;
