+++ src/types/index.ts
@@ -1,0 +1,19 @@
+export interface CardDTO {
+  id: string;
+  title: string;
+  sizes: string[];
+  basePrice: number;
+  pages: {
+    title: string;
+    templateId: string;
+  }[];
+}
+
+export interface Card {
+  getSummary: () => CardSummary;
+}
+
+export interface CardSummary {
+  id: string;
+  title: string;
+}
